+++ goldens/circular-deps/packages.json
@@ -8,10 +8,6 @@
     "packages/angular_devkit/build_angular/src/tools/esbuild/utils.ts"
   ],
   [
-    "packages/angular_devkit/build_angular/src/tools/webpack/utils/stats.ts",
-    "packages/angular_devkit/build_angular/src/utils/bundle-calculator.ts"
-  ],
-  [
     "packages/angular/cli/src/analytics/analytics-collector.ts",
     "packages/angular/cli/src/command-builder/command-module.ts"
   ],
